--- src/main/java/baseball/domain/BaseballCode.java
@@ -1,20 +1,0 @@
-package baseball.domain;
-
-import java.util.List;
-
-public final class BaseballCode {
-	private final List<Integer> codes;
-
-	public BaseballCode(List<Integer> codes) {
-		this.codes = codes;
-	}
-
-	public List<Integer> getCodes() {
-		return codes;
-	}
-
-	public BaseballCode makeNewBaseballCode(List<Integer> codes) {
-		return new BaseballCode(codes);
-	}
-
-}
